+++ src/awds/RTTMetric.h
@@ -13,7 +13,7 @@ namespace awds {
     protected:
 
 	virtual RTopology::link_quality_t my_get_quality(NodeDescr &ndescr);
-	virtual unsigned long my_calculate(RTopology::link_quality_t forward,RTopology::link_quality_t backward);
+	virtual uint32_t my_calculate(RTopology::link_quality_t forward,RTopology::link_quality_t backward);
 
 	struct s_rtt_data {
 	    gea::Duration time;
